chore(runner_action.go): extract cache related elements
